perf: remove some unnecessary code
## s3api_bucket_handlers.go
Bucket and collection are not in the same dimension. Creating a bucket does not create a corresponding collection, so it is a bit difficult to traverse the collections to determine whether a bucket exists.

## command_fs_configure.go
"check collection":
We can configure overGrowthCount, data center, rack, replication, ttl, etc. through fs_configure according to the path matching of the filer rule, which can be very convenient to do different configurations for different buckets (mapped to different paths), but now the check here collection" judgment restricts us from doing so, I don't think this judgment is necessary

"check replication":
The volumeGrowthCount judgment here is unnecessary, right?
When volumeLayout=1, volumeLayout contains a set of volumes, these volumes have satisfied our configuration (dc, rack, replication and so on), adding volumeLayout can spread the write pressure in different volume groups, but it will also increase disk usage , if required in subsequent scenarios, it can also be operated through the interface

thx!