Avoid race conditions when opening files
Avoid re-setting the same instance of filer_pb.Entry to
FileHandle.entry on every Open call.
Once a file has been opened for the first time, the entry is set and
does not change anymore.
This avoids race conditions, for example when working on the same
file from multiple threads.