Fix corrupted mount reads after changes on fh
When a file is opened multiple times - no matter if single- or multi-
threaded - and something was written, or a metadata attribute was
changed, then the following reads would always return wrong / old
data until all file handles were closed and the file was reopened.
The reason was that the outdated reader was not reset properly with a
fresh chunk view.