weed/shell: Fix comparison of volume capacity
The `volume.balance` command failed to properly balance volumes, because when
it sorted nodes by capacity, it used this comparison function:

```golang
func(a, b *Node) int {
  return int(a.localVolumeRatio(capacityFunc) - b.localVolumeRatio(capacityFunc))
})
```

This doesn't work because `int(-0.1)` is zero, not a negative number, so nodes
were not properly sorted.

This commit makes use of `cmp.Compare()` instead, which returns the expected
result.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>