Fix deadlock in lock table locks
This fixes a race condition in which multiple Goroutines release and
acquire locks for the same file. Because the lock entries in the
lock table and the counters in the lock entries themselves are
independent and not synchronized, it can happen that two Goroutines
acquire a lock and a third Goroutine releases the same lock in
between them, leading to two lock entries being created. Those two
lock entries would now wait forever on entry.cond.Wait() and can't
proceed. The fix is to track the locks that are in flight in the
lock table itself, so that it knows when it is safe to release a
lock before creating a second instance.