add http endpoint to get the size of a collection
get the size of a collection without using the cli is useful in a
context where we use seaweedfs as a microservice and we want to meter
the fs usage for a given user, that writes in a single s3 bucket. Using
s3 api is slow because we have to loop over all files. So this endpoint
comes very useful to get directly from seaweedfs the total usage with a
single call.

The endpoint is /collection/info with POST form fields:
 - collection
 - detail (if 'true' provides details for each volume, otherwise only
   the total)