Remove "Content-Length" header if http.Error is issued
The code which attempts to send the data, sets Content-Length
to expected values. But if there were an error and we issue
http.Error the Content-Length header should be reset to the error
message length.

This should be unnesesary, since it is already done
net/http package, see
https://cs.opensource.google/go/go/+/refs/tags/go1.23.0:src/net/http/server.go;drc=773767def0e0f29584a69bd760430167b7479d7d;l=2245

But maybe SeaweedFS uses an earlier version.