What problem are we solving? Fix: #6579
How are we solving the problem?
In the Helm chart, we remove volume and volumeMounts for the certificates that are not used.

How is the PR tested?
We ran SeaweedFS in a Kubernetes Cluster with a joint Filer and S3 server in one container, with leveldb2 as the filer storage.

Full communication between the various components stays intact, despite the removal of the unused certificates.

Checks
[ ] I have added unit tests if possible.
[ ] I will add related wiki document changes and link to this PR after merging.